Shift panel-related code to its own class
